New dark theme.
