+++ app/src/main/java/jp/projectoffline/todo/MainActivity.java
@@ -1,24 +1,17 @@
 package jp.projectoffline.todo;
 
-import android.app.AlertDialog;
-import android.app.Dialog;
 import android.content.ContentValues;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
-import android.text.InputFilter;
-import android.text.InputType;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
 import android.widget.ListView;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.DialogFragment;
@@ -55,7 +48,12 @@ public class MainActivity extends AppCompatActivity {
         final SQLiteDatabase db = helper.getWritableDatabase();
         //arrayItemsに追加された要素をリストで表示
         arrayItems = new ArrayList<>();
-        final ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        final ArrayAdapter<String> adapter;
+        if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("dark_theme", true)){
+            adapter = new ArrayAdapter<>(this, R.layout.text_color, arrayItems);
+        } else {
+              adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        }
         listView = findViewById(R.id.list_view);
         listView.setAdapter(adapter);
         //要素の削除
@@ -120,7 +118,13 @@ public class MainActivity extends AppCompatActivity {
     public void deleteToDo(){
         final SQLiteDatabase db = helper.getWritableDatabase();
 
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        arrayItems = new ArrayList<>();
+        ArrayAdapter<String> adapter;
+        if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("dark_theme", true)){
+            adapter = new ArrayAdapter<>(this, R.layout.text_color, arrayItems);
+        } else {
+            adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        }
         listView.setAdapter(adapter);
 
         arrayItems.clear();
@@ -149,7 +153,13 @@ public class MainActivity extends AppCompatActivity {
         //データベースに値を追加
         db.insert("todo_items", null, values);
 
-        ArrayAdapter<String> adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        arrayItems = new ArrayList<>();
+        ArrayAdapter<String> adapter;
+        if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("dark_theme", true)){
+            adapter = new ArrayAdapter<>(this, R.layout.text_color, arrayItems);
+        } else {
+            adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, arrayItems);
+        }
         listView.setAdapter(adapter);
         //リストを初期化して更新
         arrayItems.clear();
@@ -166,72 +176,7 @@ public class MainActivity extends AppCompatActivity {
         adapter.notifyDataSetChanged();
         c.close();
     }
-    //削除時のfragment
-    public static class DeleteDialogFragment extends DialogFragment {
-        @Override
-        @NonNull
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
-            alert.setMessage(R.string.confirm);
-            alert.setPositiveButton(R.string.dialog_positive2, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialogInterface, int i) {
-                    //MainActivityの関数と連携
-                    MainActivity mainActivity = (MainActivity) getActivity();
-                    mainActivity.deleteToDo();
-                }
-            });
-            alert.setNeutralButton(R.string.dialog_neutral, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialogInterface, int i) {
 
-                }
-            });
-            //ダイアログを表示
-            return alert.create();
-        }
-    }
-    //入力時のfragment
-    public static class AlertDialogFragment extends DialogFragment {
-
-        @Override
-        @NonNull
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            final EditText editText = new EditText(getActivity());
-            //入力制限 (1行, 256文字まで)
-            InputFilter[] inputFilter = new InputFilter[1];
-            inputFilter[0] = new InputFilter.LengthFilter(256);
-
-            editText.setFilters(inputFilter);
-            editText.setInputType(InputType.TYPE_CLASS_TEXT);
-
-            AlertDialog.Builder alert = new AlertDialog.Builder(getActivity());
-            alert.setView(editText);
-            alert.setTitle(R.string.input);
-
-            alert.setPositiveButton(R.string.dialog_positive1, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialogInterface, int i) {
-                    //入力した文字を取得
-                    String str = editText.getText().toString();
-                    //MainActivityの関数と連携
-                    MainActivity mainActivity = (MainActivity) getActivity();
-                    //空白であればデータベースに追加しない
-                    if (!str.equals("")) {
-                        mainActivity.addToDo(str);
-                    }
-                }
-            });
-            alert.setNeutralButton(R.string.dialog_neutral, new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialogInterface, int i) {
-                    //Cancel
-                }
-            });
-            //ダイアログを表示
-            return alert.create();
-        }
-    }
     @Override
     protected void onDestroy() {
         super.onDestroy();
