+++ app/src/main/res/values/colors.xml
@@ -3,8 +3,8 @@
     <color name="colorPrimary">#00C853</color>
     <color name="colorPrimaryDark">#43A047</color>
     <color name="colorAccent">#D81B60</color>
-    <color name="black">#424242</color>
+    <color name="black">#000000</color>
+    <color name="grey">#424242</color>
     <color name="white">#FFFFFF</color>
     <color name="textColor">#757575</color>
-    <color name="listColor">#43A047</color>
 </resources>
